requirements.txt
streamlit>=1.32.0
pandas==2.0.3
ortools==9.7.2996
google-generativeai==0.3.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
africastalking==2.0.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==0.8.11
openpyxl==3.1.2
tabulate==0.9.0
numpy
google-cloud-firestore
firebase-admin
requests
streamlit-autorefresh
pydantic<2
PyPDF2>=3.0
python-docx>=0.8.11
 # optional